Split related articles into own file, simplify mapping logic
